+++ packages/dht/test/integration/WebsocketClientConnectorRpc.test.ts
@@ -32,11 +32,11 @@ describe('WebsocketClientConnectorRpc', () => {
             mockWebsocketClientConnectorRpc.requestConnection
         )
 
-        rpcCommunicator1.on('outgoingMessage', (message: RpcMessage) => {
+        rpcCommunicator1.setOutgoingMessageListener(async (message: RpcMessage) => {
             rpcCommunicator2.handleIncomingMessage(message)
         })
 
-        rpcCommunicator2.on('outgoingMessage', (message: RpcMessage) => {
+        rpcCommunicator2.setOutgoingMessageListener(async (message: RpcMessage) => {
             rpcCommunicator1.handleIncomingMessage(message)
         })