+++ packages/sdk/src/Authentication.ts
@@ -1,9 +1,9 @@
-import { BrowserProvider, AbstractSigner, Provider, computeAddress, Wallet } from 'ethers'
-import { EthereumAddress, hexToBinary, toEthereumAddress, wait, createSignature } from '@streamr/utils'
+import { createSignature, hexToBinary, toEthereumAddress, UserID, wait } from '@streamr/utils'
+import { AbstractSigner, BrowserProvider, computeAddress, Provider, Wallet } from 'ethers'
 import pMemoize from 'p-memoize'
 import { PrivateKeyAuthConfig, ProviderAuthConfig, StrictStreamrClientConfig } from './Config'
-import { pLimitFn } from './utils/promises'
 import { RpcProviderSource } from './RpcProviderSource'
+import { pLimitFn } from './utils/promises'
 
 export const AuthenticationInjectionToken = Symbol('Authentication')
 
@@ -11,7 +11,7 @@ export type SignerWithProvider = AbstractSigner<Provider>
 
 export interface Authentication {
     // always in lowercase
-    getAddress: () => Promise<EthereumAddress>
+    getAddress: () => Promise<UserID>
     createMessageSignature: (payload: Uint8Array) => Promise<Uint8Array>
     getTransactionSigner: (rpcProviderSource: RpcProviderSource) => Promise<SignerWithProvider>
 }