+++ packages/sdk/src/Message.ts
@@ -1,4 +1,4 @@
-import { EthereumAddress, StreamID } from '@streamr/utils'
+import { StreamID, UserID } from '@streamr/utils'
 import { SignatureType, StreamMessage } from './protocol/StreamMessage'
 
 /**
@@ -45,7 +45,7 @@ export interface Message {
     /**
      * Publisher of message.
      */
-    publisherId: EthereumAddress
+    publisherId: UserID
 
     /**
      * Identifies the message chain the message was published to.