+++ packages/sdk/src/utils/persistence/ServerPersistence.ts
@@ -37,13 +37,13 @@ export default class ServerPersistence implements PersistenceContext {
 
     private constructor({
         loggerFactory,
-        clientId,
+        ownerId,
         migrationsPath,
         onInit
     }: ServerPersistenceOptions) {
         this.logger = loggerFactory.createLogger(module)
         const paths = envPaths('streamr-sdk')
-        this.dbFilePath = resolve(paths.data, join('./', clientId, `GroupKeys.db`))
+        this.dbFilePath = resolve(paths.data, join('./', ownerId, `GroupKeys.db`))
         this.migrationsPath = migrationsPath
         this.onInit = onInit
         this.init = pOnce(this.init.bind(this))