+++ CHANGELOG.md
@@ -18,7 +18,11 @@ Changes before Tatum release are not documented in this file.
 
 #### Removed
 
+- Remove support for legacy encryption keys (https://github.com/streamr-dev/network/pull/2757)
+
 #### Fixed
+
+- Fixed flag expiration time in `Operator#getExpiredFlags` (https://github.com/streamr-dev/network/pull/2739)
 
 #### Security